[Editorial] use consistent hyphen, volerun --> vole-run
Suspected publisher typo.

This one is tricky.  The text has 1 "vole-runs" and 1 "volerun"

MW dictionary has no entry for "vole-run", "vole run", or "volerun.

Google N-gram viewer shows "vole run" as the only variant to show up.

https://books.google.com/ngrams/graph?content=vole-runs%2Cvoleruns%2Cvole+runs&year_start=1800&year_end=2019&corpus=en-GB&smoothing=3&case_insensitive=false

P3 Proofreader suggested changing "volerun" to "vole run".  Which the PP did, but didn't change both "vole-runs" and "volerun" to "vole run(s)"

Alternately, all of the editions at IA use either "vole-run" or "vole-*run"

https://archive.org/search?query=%28%28vole-run%29+OR+%28volerun%29+OR+%28%22vole+run%22%29%29+AND+title%3A%28Tarka+otter%29+AND+mediatype%3A%28texts%29&sin=TXT&sort=-date

So, I think there is merit to (A) leaving "vole-runs" and "volerun" exactly as they are in the page scans, (B) modernizing both instances to N-gram viewer "vole run(s), or (C) updating "volerun" to "vole-run" to agree with all the other editions of the book.

I'm suggesting we go with option C, because the prevalence of "vole-run" in all the other editions makes me think that "volerun" was likely a printers typo.