Fix typo, should be "water-lies"
Original text
>> Near her was Captain, a black-and-tan rough dog, who looked like a lurcher; the huntsman did not take Captain to important meets, for Captain’s voice was like a knife whose edge is turned. He did not throw his tongue, he screamed; and sometimes in his excitement he babbled, flinging water-lies about.

PP changed "water-lies" to "waterflies".  But I don't think this is correct given the context.  He is describing what a dog does when he gets excited.  I'm not familiar with the Devon dialect of English, and I don't know that "water-lies" means, but I suspect, given the context, that it it referring to droplets of saliva (spittle) that get thrown around when a dog is excited and making a lot of noise.

Furthermore, none of the 23 editions avaliable at IA substitute this change.  One edition substitutes the word "water-lilies", but the rest keep "water-lies".

https://archive.org/search?query=%28%22excitement+he+babbled%22%29+AND+title%3A%28Tarka+otter%29+AND+mediatype%3A%28texts%29&sin=TXT&sort=-date