Fix typo, should be "pobbles"
Page scans show:
>> the stone setts and pobbles of the wall’s apron, whose cracks held the little ruddy winter leaves of the sea-beet,

Other editions treat this word inconsistently,  of the 23 uploads to IA, 12 keep "pobbles" and 11 change it to "pebbles"

https://archive.org/search?query=%28%22stone+setts+and%22%29+AND+title%3A%28Tarka+otter%29+AND+mediatype%3A%28texts%29&sin=TXT&sort=-date

"Pobbles" is not it the MW dictionary, it lists "pebbles" and "cobbles" as the top two suggestions.  Google Ngram shows that both "pebbles" and "cobbles" are much more frequent words than "pobbles".

I believe that "pobbles" is a word invented by the author.  He is known to invent words (see source in previous commit about "twired").

I found one blogger that comments on the phrase "setts and pobbles."  He writes:

>> Couldn’t find this ‘pobble’ in the 600,000 words of the OED, until I realized it must be referring to ‘setts and cobbles’.   A ‘sett’ is a cobblestone and the two are often mentioned together.

>> I prefer ‘pobble’ which must be a pebble-cobbled cobblestone.  Or a fortuitous typo?

https://www.writingredux.com/words-setts-and-pobbles/