EOL breast-*bone --> breastbone
There is only one instance of the word within the book. However there are similar compound words ending in "bone" as follows:

2 "backbone"
1 "jawbone"
1 "leg-bones"
1 each "fishbone" "fish-*bones" "fish-bones" and "fishbones"

MW dictionary autocorrects "breast-bone" to "breastbone".

https://www.merriam-webster.com/dictionary/breastbone

Google N-gram viewer shows that "breastbone" is more popular.

https://books.google.com/ngrams/graph?content=breast-bone%2Cbreastbone&year_start=1800&year_end=2019&corpus=en-GB&smoothing=3

On the other hand, IA shows that "breast-bone" is more common in other editions.

https://archive.org/search?query=%28%22and+flesh+together%22%29+AND+title%3A%28Tarka+otter%29+AND+mediatype%3A%28texts%29&sin=TXT&sort=-date

So, the information is somewhat contracitory here. I think we should go with "breastbone" because it is similar to the author's use of "backbone" and "jawbone" (but not "leg-bones") and because the Ngram viewer shows "breastbone" to be more prevalent.